Move all the survey's data to the right folder.
