+++ carrerait-roadmap/src/app/data/itWorkerConditionalResponses.ts
